requirements/base.txt
pytz==2021.3                             # https://github.com/stub42/pytz
python-slugify==5.0.2                    # https://github.com/un33k/python-slugify
python-dateutil==2.8.2                   # https://github.com/dateutil/dateutil
Pillow==8.4.0                            # https://github.com/python-pillow/Pillow
argon2-cffi==21.1.0                      # https://github.com/hynek/argon2_cffi
whitenoise==5.3.0                        # https://github.com/evansd/whitenoise
redis==3.5.3                             # https://github.com/andymccurdy/redis-py
hiredis==2.0.0                           # https://github.com/redis/hiredis-py
celery==5.2.3                            # https://github.com/celery/celery
django-celery-beat==2.2.1                # https://github.com/celery/django-celery-beat
flower==1.0.0                            # https://github.com/mher/flower
beautifulsoup4==4.10.0                   # https://pypi.org/project/beautifulsoup4
PyYAML==5.4.1                            # https://github.com/yaml/pyyaml
PyJWT==2.3.0                             # https://github.com/jpadilla/pyjwt
botocore==1.21.65                        # https://github.com/boto/botocore
xmltodict==0.12.0                        # https://github.com/martinblech/xmltodict
ipwhois==1.2.0                           # https://github.com/secynic/ipwhois
argparse==1.4.0                          # https://github.com/ThomasWaldmann/argparse/
lxml==4.6.4                              # https://github.com/lxml/lxml
bcrypt==3.2.0                            # https://github.com/pyca/bcrypt
cryptography==36.0.0                     # https://github.com/pyca/cryptography

future==0.18.2
arrow==1.2.2
backports.zoneinfo==0.2.1
dateparser==1.1.1
moment==0.12.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
times==0.7
tzdata==2022.1
tzlocal==4.2

requests==2.26.0                         # https://github.com/psf/requests
gevent==21.8.0                           # https://github.com/gevent/gevent
tldextract==3.1.2                        # https://github.com/john-kurkowski/tldextract
python-http-client==3.3.4                # https://github.com/sendgrid/python-http-client
validate-email==1.3                      # https://github.com/syrusakbary/validate_email
deepdiff==5.6.0                          # https://github.com/seperman/deepdiff
ipinfo==4.2.1                            # https://github.com/ipinfo/python
jsonpath-ng==1.5.3                       # https://github.com/h2non/jsonpath-ng
requests-toolbelt==0.9.1                 # https://github.com/requests/toolbelt

watchgod==0.7                            # https://github.com/samuelcolvin/watchgod
uvicorn[standard]==0.15.0                # https://github.com/encode/uvicorn

cryptography==36.0.0                     # https://github.com/pyca/cryptography

# Django
# ------------------------------------------------------------------------------
django==3.2.9                            # https://github.com/django/django
django-environ==0.8.1                    # https://github.com/joke2k/django-environ
django-model-utils==4.2.0                # https://github.com/jazzband/django-model-utils
django-rest-auth==0.9.5                  # https://github.com/Tivix/django-rest-auth
django-redis==5.1.0                      # https://github.com/jazzband/django-redis
django-simple-history==3.0.0             # https://github.com/treyhunner/django-simple-history
django-cacheops==6.0                     # https://github.com/Suor/django-cacheops
django-cors-headers==3.10.1              # https://github.com/ottoyiu/django-cors-headers
django-ratelimit==3.0.1                  # https://github.com/jsocol/django-ratelimit
django-crispy-forms==1.13.0              # https://github.com/django-crispy-forms/django-crispy-forms
django-tables2==2.4.1                    # https://github.com/jieter/django-tables2
django-oauth-toolkit==1.5.0              # https://github.com/jazzband/django-oauth-toolkit
django-json-widget==1.1.1

# Django REST Framework
djangorestframework==3.12.4              # https://github.com/encode/django-rest-framework
drf-jwt==1.19.1                          # https://github.com/Styria-Digital/django-rest-framework-jwt/
django-filter==21.1                      # https://github.com/carltongibson/django-filter

drf-yasg==1.20.0                         # https://github.com/axnsan12/drf-yasg
psycopg2==2.9.2                          # https://github.com/psycopg/psycopg2
django-rest-auth[with_social]
drf-social-oauth2
